Refactor NMEADisplay into cards
