+++ Makefile
@@ -1,4 +1,7 @@
 
+dev:
+	npm run dev
+
 build:
 	npm run build
