+++ examples/nmea-demo/Makefile
@@ -1,12 +1,16 @@
 
-dev:
+dev: node_modules
 	npm run dev
 
-build:
+build: node_modules
 	npm run build
 
-simple:
+simple: node_modules
 	rm -r node_modules/nmea-simple
 	npm install nmea-simple@file:/home/john/src/nmea-simple --force 
 	npm run build
 	rm -r node_modules/.vite/deps/
+
+node_modules:
+	npm install
+
